DB-1504: CR Revisions
