+++ starters/gatsby-wordpress-starter/package.json
@@ -26,10 +26,6 @@
 		"coverage": "vitest run --coverage"
 	},
 	"dependencies": {
-		"@testing-library/react": "12.1.5",
-		"@vitejs/plugin-react": "^2.1.0",
-		"autoprefixer": "^10.4.12",
-		"esm": "^3.2.25",
 		"gatsby": "^4.24.0",
 		"gatsby-plugin-image": "^2.24.0",
 		"gatsby-plugin-manifest": "^4.24.0",
@@ -43,7 +39,6 @@
 		"graphql": "^15.0.0",
 		"lodash": "^4.17.21",
 		"mitt": "^3.0.0",
-		"postcss": "^8.4.16",
 		"react": "17.0.2",
 		"react-dom": "17.0.2",
 		"typeface-merriweather": "1.1.13",
@@ -53,8 +48,12 @@
 		"@babel/core": "^7.19.3",
 		"@pantheon-systems/wordpress-kit": "^2.8.0-canary.0",
 		"@tailwindcss/typography": "^0.5.7",
+		"@testing-library/react": "12.1.5",
+		"@vitejs/plugin-react": "^2.1.0",
 		"@vitest/coverage-c8": "^0.24.1",
+		"autoprefixer": "^10.4.12",
 		"dumper.js": "^1.3.1",
+		"postcss": "^8.4.16",
 		"prettier": "^2.7.1",
 		"tailwindcss": "^3.1.8",
 		"vitest": "^0.23.4",
